refactor: rename InjuryStatus class to Injury
